+++ packer/provisioners/base-provisioning.sh
@@ -4,11 +4,29 @@ application_file_path="/vagrant/installed-application.md"
 
 # install docker
 sudo apt update
-sudo apt install -y docker.io
+sudo apt-get -y install \
+    apt-transport-https \
+    ca-certificates \
+    curl \
+    gnupg-agent \
+    software-properties-common
+
+# add Docker’s official GPG key
+curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+
+# add docker repository
+sudo add-apt-repository \
+   "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
+   $(lsb_release -cs) \
+   stable"
+
+# install docker
+sudo apt-get update
+sudo apt-get -y install docker-ce docker-ce-cli containerd.io
 
 # install docker service
-sudo systemctl start docker
-sudo systemctl enable docker
+# sudo systemctl start docker
+# sudo systemctl enable docker
 
 # add vagrant user to docker group
 sudo groupadd docker